rename ActionObject to Action and add standard interfaces
